Add semantics to service nodes
Previously, it was unkown whether nodes such as ApiPayloadSecurity,
or any other Node for that matter, was a candidate for GetSingleton()
or GetScoped().  Now, those methods can only accept a template parameter
of type ServiceNode, and Nodes meant to be used as such must inherit
from it, eliminating the ambiguity as to a node's candidacy and usage.