Bugfix: Use First() and exception handling to avoid trying to remove nonexistent clients
Just because loggedInPlayer was of type int? doesn't mean that FirstOrDefault() was
going to be able to return null, because the input type was a collection of regular
integers, not nullable integers.  The default value is 0, so previously, when the
FindOrDefault() predicate failed, the code would attempt to disconnect peer 0 and
remove key 0 from _clients and _players, none of which exist.  Instead, allowing
Find() to throw an exception when the predicate fails allows us to escape from
the rest of the block.