AES encrypt MP login payload with RSA secured key
Previously, purely RSA was used to encrypt/decrypt user login credentials.
As a result, user credentials were subject to hard length limits, where
exceeding the block size of the RSA engine would throw an out of bounds
exception.  The RSA algorithm by itself is not well suited to arbitrary-
length payload encryption.  Enter AES-CBC, which can do this job.  It
requires an additional key of its own, to be generated by the sender (the
MP client), yet since it's symmetrical it must be asymmetrically secured,
which is the new role of the API's transient key pair.