Use WSS and HTTPS conditionally
Most importantly, the server will load and use a certificate if one
is found at ./x509/sharpscape.crt.  This will automatically cause
the server to listen over Secure WebSocket.  However, there is currently
no way to automatically switch the client's listening protocol. The
user needs to know ahead of time whether the server is listening on a
secured websocket or not.