Clients can mostly synchronize now
The clients emulate their own copies of the world, but the server
does not yet.  It needs to, in order to be authoritative over the
positions of players, and to handle validity checking for movement
and in general will lay the groundwork for more complex player
interactions that can be validated by a central authroity.