Bugfix: Free NetworkServiceNode on World unload
In some cases, logging out of a client and switching accounts would leave
ghost avatars in memory from the previous login.  When the world is unloaded,
the session is over, so all memory of network objects should be freed.  A new
network service node will be instantiated automatically when needed.